Require explicit version-release of chromium-libs from media-freeworld
Since the freeworld package is built in another repository,
it may be out of sync. This makes sure `dnf update` won't just
remove the installed version of chromium in order to update
chromium-libs-media-freeworld if that happens to be in the repo
sooner than chromium.

Reported as #4572 in that other repository.